+++ apps/app/src/app/@common/transaction-entry/ui/TransactionEntryList/Item/type/Gallery/Gallery.tsx
@@ -1,19 +1,82 @@
 import type { Container } from "@use-pico/client/ui/container";
+import { Typo } from "@use-pico/client/ui/typo";
+import { withFallback } from "@use-pico/client/utils";
+import { toTimeDiff } from "@use-pico/common/time";
 import type { tTransactionEntryGallery } from "@zbav-se.me/sdk/api/user";
-import { type FC, Suspense } from "react";
-import { Data } from "./Data";
+import { withTransactionEntryGalleryFetchQuery } from "@zbav-se.me/sdk/query/user/transaction-entry";
+import { HeroImage } from "@zbav-se.me/ui/img";
+import { useLocale } from "@use-pico/client/hook";
+import type { MarkSuspense } from "@use-pico/client/type";
+import { useState } from "react";
+import { useUpload } from "~/app/@common/gallery/hook/useUpload";
+import { GalleryPreviewSheet } from "~/app/@common/gallery/ui/GalleryPreviewSheet";
 import { Pending } from "./Pending";
+import { TypeContainer } from "../TypeContainer";
 
 export namespace Gallery {
-	export interface Props extends Container.Props {
+	export interface Props extends Container.Props, MarkSuspense.Props {
 		transactionEntry: tTransactionEntryGallery;
 	}
 }
 
-export const Gallery: FC<Gallery.Props> = (props) => {
+export const Gallery = withFallback(({ _suspense, transactionEntry, ...props }: Gallery.Props) => {
+	const locale = useLocale();
+	const { data: gallery } = withTransactionEntryGalleryFetchQuery.useSuspenseQuery({
+		where: {
+			transactionEntryId: transactionEntry.id,
+		},
+	});
+	const hero = useUpload(gallery.items);
+	const [isGalleryOpen, setIsGalleryOpen] = useState(false);
+
 	return (
-		<Suspense fallback={<Pending transactionEntry={props.transactionEntry} />}>
-			<Data {...props} />
-		</Suspense>
+		<TypeContainer
+			direction={transactionEntry.direction}
+			ui={{
+				position: "relative",
+				theme: undefined,
+				background: undefined,
+				round: undefined,
+				inner: undefined,
+			}}
+			className={"h-48"}
+			{...props}
+		>
+			<HeroImage
+				src={hero.url}
+				visible
+				ui={{
+					tone: "neutral",
+					theme: "light",
+					background: "default",
+					round: "default",
+				}}
+				onClick={() => setIsGalleryOpen((prev) => !prev)}
+			/>
+
+			<Typo
+				label={toTimeDiff({
+					locale,
+					time: transactionEntry.createdAt,
+					type: "relative",
+				})}
+				ui={{
+					tone: "neutral",
+					theme: "light",
+					background: "default",
+					text: "sm",
+					round: "default",
+					inner: "default",
+					opacity: "8",
+					snapTo: "bottom-left",
+				}}
+			/>
+
+			<GalleryPreviewSheet
+				uploads={gallery.items.map((item) => item.upload)}
+				isOpen={isGalleryOpen}
+				onClose={() => setIsGalleryOpen(false)}
+			/>
+		</TypeContainer>
 	);
-};
+}, Pending);
